Update README.md
Clarify what flushing interval means (so it can be compared to reporting interval in T2 SDKs).

I didn't change this, but I wondered why the comment for the proxy client's `flushIntervalSeconds` says the default is 5, while the default for the direct ingestion client is 1. Note that the doc for `pushFlushInterval` in the `wavefront.conf` file says the default is 1 second. Should the 5 be changed to 1 here?